requirements.txt
pandas>=2.0.0
matplotlib>=3.8.0
chromadb>=0.4.22
sentence-transformers>=2.2.2
google-generativeai>=0.3.0
python-dotenv>=1.0.0
PyQt6>=6.5.0
pyqtdarktheme>=0.1.7
Pillow>=10.0.0
requests>=2.31.0
pyinstaller>=6.0.0
